Trick the Mozilla linter into ignoring the states.json
Otherwise it complains that it's too large and demands we take the
data and put it in... a json file X_X

I tried modifying the generator to split files (1 file per state) but when the directory
was at 1GB and not done (metadata? I dunno) I gave up.
Maybe a less granular split would work? I don't see an advantage over this
.bin trick though.